+++ package.json
@@ -26,10 +26,10 @@
     "build:clean": "rimraf ./build",
     "build:dll": "babel-node ./internals/scripts/dependencies.js",
     "build": "npm run build:client && npm run build:server",
-    "start": "cross-env NODE_ENV=development babel-node app/server",
+    "start": "cross-env NODE_ENV=development node app/server",
     "start:tunnel": "cross-env NODE_ENV=development ENABLE_TUNNEL=true babel-node app/server",
     "start:production": "npm run test && npm run build && npm run start:prod",
-    "start:prod": "cross-env NODE_ENV=production babel-node build/server/server",
+    "start:prod": "cross-env NODE_ENV=production node build/server/server",
     "presetup": "npm i chalk shelljs",
     "setup": "babel-node ./internals/scripts/setup.js",
     "postsetup": "npm run build:dll",
@@ -145,6 +145,7 @@
     "cross-env": "5.2.0",
     "dotenv": "^6.0.0",
     "enhanced-resolve": "^4.1.0",
+    "esm": "^3.1.0",
     "express": "4.16.3",
     "fontfaceobserver": "2.0.13",
     "graphql": "^14.0.2",
